pyproject.toml
[project]
name = "agentgenius"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "duckduckgo-search>=7.2.1",
    "logfire>=2.11.1",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pip>=24.3.1",
    "psutils>=3.3.8",
    "pydantic-ai>=0.0.15",
    "pydub>=0.25.1",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "setuptools>=75.8.0",
    "simpleaudio>=1.0.4",
    "streamlit>=1.41.1",
]

[tool.ruff]
line-length = 120
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
lint.ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip() without an explicit strict= parameter
]

[tool.dependencies]
nest-asyncio = "*"
